Finalizado el esqueleto del modo runner del planeta fuegoHielo
Añadido un placeholder para el combate contra la hydra

Añadido puzzle planetario en la última galaxia.

+++ Assets/Scripts/Gravity and Planets/DieOnTouch.cs
@@ -7,13 +7,13 @@ public class DieOnTouch : MonoBehaviour {
 	public bool trigger = true;
 
 	void OnCollisionEnter (Collision collider) {
-		if(collider.gameObject.tag == "Player" && collision){
+		if(collider.gameObject.tag == "Player" && collision && isActiveAndEnabled){
 			GameManager.playerController.kill();
 		}
 	}
 
 	void OnTriggerEnter (Collider collider) {
-		if(collider.gameObject.tag == "Player" && trigger){
+		if(collider.gameObject.tag == "Player" && trigger && isActiveAndEnabled){
 			GameManager.playerController.kill();
 		}
 	}
